rework system connection

+++ javaLab3/src/main/java/org/example/connection/MessageIdE.java
@@ -17,7 +17,7 @@ public enum MessageIdE {
         this.value = value;
     }
 
-    public int getValue() {
+    public byte getValue() {
         return value;
     }
 }
